+++ HISTORY.md
@@ -1,3 +1,10 @@
+## 1.1.0 (2011-11-22)
+
+* new `FILE` sub-commands `USER_NAME`, `GROUP_NAME` and `PERMISSIONS` (UNIX only)
+* new command `current_timestamp`
+* new command `get_timestamp_component`
+* bug fixes for Windows and Cygwin
+
 ## 1.0.0 (2011-07-02)
 
 * First release